cri/sandbox: autodect mutable stage1 capabilities
This documents and autodetects stage1 mutable cabability detection. It
is needed for stage1 images which do not support mutable operations, and
retains backwards compatibility with old stage1 images.
